chore(#3777): connection query cleanup | extract resource loader
## Summary

Extracts `ConnectionResourceLoader` from `ConnectionQuery` as part of the cleanup initiative (#3756).

### Changes

- **Created `ConnectionResourceLoader.cs`** — new focused class in `Queries/Connection/` containing:
  - `LoadResourcesByKeyAsync` — loads resources per rightholder assignment
  - `LoadInstancesByKeyAsync` — loads instances per rightholder assignment
  - `GetRightholderAssignments` (static) — filters keys to rightholder assignments with no mutable state

- **Simplified `ConnectionQuery.cs`**:
  - Removed mutable `_rightholderAssignments` field and its reset logic
  - Removed the three extracted private methods (~150 lines)
  - Delegates to `ConnectionResourceLoader` via a `_resourceLoader` field

### Constraints met

- ✅ No change in returned resources or instances
- ✅ `ConnectionQuery` public API unchanged
- ✅ All 305 integration tests pass

Closes #3776
